requirements.txt
numpy>=2.0.1
pyserial>=3.5
rerun-sdk>=0.15.0